requirements.txt
Flask
sqla-wrapper
beautifulsoup4
requests
mysql-connector-python
sqlalchemy 
